# internal/handler/openapi.yaml
openapi: 3.0.3
info:
  title: tileserve-go
  description: >
    A small Go HTTP server for managing maps and their versioned tile uploads,
    backed by Postgres and secured with JWT authentication. Maps have
    versioned zip uploads extracted to `<data-root>/<uuid>/<version>/` on
    disk, served back at `/maps/{id}/version/{version}/...`.
  version: "1.0.0"
servers:
  - url: http://localhost:8085
...
security:
  - bearerAuth: []
  - tokenQuery: []
paths:
...
  /groups:
    get:
      tags: [groups]
      summary: List groups
      description: >-
        Open to any authenticated user (e.g. so a map owner can pick a
        group to grant a per-map permission to).
      parameters:
        - name: search
          in: query
          description: Case-insensitive substring match against name.
          schema:
            type: string
      responses:
        "200":
          description: Array of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [groups]
      summary: Create a group
      description: Requires the acting user's `is_admin` permission.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupRequest"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          description: Invalid request body, or missing name
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Group name already taken
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /groups/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [groups]
      summary: Get a single group
      description: >-
        Open to any authenticated user, same rationale as `GET /groups`.
      responses:
        "200":
          description: The group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Group not found
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
    put:
      tags: [groups]
      summary: Update a group's name and permission bundle
      description: >-
        Requires the acting user's `is_admin` permission. All fields are
        replaced (not partial).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupRequest"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          description: Invalid request body, or missing name
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Group not found
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "409":
          description: Group name already taken
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
    delete:
      tags: [groups]
      summary: Delete a group
      description: >-
        Requires the acting user's `is_admin` permission. Its memberships and
        per-map grants are removed along with it.
      responses:
        "204":
          description: Group deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Group not found
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Every bearer token is a JWT, of one of two kinds: a human login/
        refresh token (`POST /login`, `POST /refresh`), signed HS256 with a
        server-wide shared secret and carrying no `kid` header; or an API
        key token, signed RS256 by the caller with the private half of a
        key pair whose public half was registered via `POST
        /users/{username}/api-keys`, carrying a `kid` header naming that
        key's id. API key tokens must be short-lived (no more than 15
        minutes between `iat` and `exp`) — the caller mints and signs a
        fresh one for each use rather than reusing a single long-lived
        token.
    tokenQuery:
      type: apiKey
      in: query
      name: token
      description: JWT (login or API key) passed as a query parameter instead of the Authorization header.
...
  responses:
    Unauthorized:
      description: Missing, invalid, or expired token
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
    Forbidden:
      description: Acting user lacks the required permission
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
...
  schemas:
    ErrorMessage:
      type: string
      description: Plain-text error message.
...
    GroupRequest:
      type: object
      required: [name, canCreate, canEdit, canDelete, isAdmin]
      properties:
        name:
          type: string
        canCreate:
          type: boolean
        canEdit:
          type: boolean
        canDelete:
          type: boolean
        canEditGeoObjects:
          type: boolean
        canDeleteGeoObjects:
          type: boolean
        canViewAll:
          type: boolean
        isAdmin:
          type: boolean
        ldapGroupDn:
          type: string
          description: >-
            The LDAP directory group (memberOf value) whose members are
            synced into this group on every login. Empty if this group isn't
            LDAP-linked.
        oidcGroupClaim:
          type: string
          description: >-
            The OIDC ID token "groups" claim value whose holders are synced
            into this group on every login. Empty if this group isn't
            OIDC-linked.
    Group:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        canCreate:
          type: boolean
        canEdit:
          type: boolean
        canDelete:
          type: boolean
        canEditGeoObjects:
          type: boolean
        canDeleteGeoObjects:
          type: boolean
        canViewAll:
          type: boolean
        isAdmin:
          type: boolean
        ldapGroupDn:
          type: string
        oidcGroupClaim:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedBy:
          type: string
...